rename TruthCatalogWriter to TruthCatalogUtils
